- OCCore Sync Engine "0.1" 	- new infrastructure to 		- schedule sync records 		- process sync records 		- handle the results, based on the Sync Journal 	- new classes OCCoreSyncRoute, OCCoreSyncParameterSet 	- reorganized internals 	- added deleteItem implementation (+ test) 	- overhauled OCSyncRecord 	- overhauled sync anchor support to also cover file-level changes (WIP) - OCConnection 	- added deleteItem implementation 	- extended deleteItem API to allow making it conditional - OCConnectionIssue 	- added multiple choice type and support methods 	- added OCConnectionIssueChoice to encapsulate choices - OCDatabase 	- added sync record support (add/update/remove/retrieve) 	- sync record support transparently stores/restores NSProgress and result handlers for sync records 	- version 4 of the metaData scheme, to support file-level sync anchor support 	- version 2 of the Sync Journal DB scheme - OCSQLite 	- support for ordering results via SQL ORDER syntax 	- support for operator other than equal (=) and more flexible query composition using new OCSQLiteQueryCondition - OCItem 	- can now also track if it has been removed - New HTTP status codes - New error codes

+++ ownCloudSDK/Vaults/Database/OCDatabase+Schemas.m
@@ -20,6 +20,17 @@
 #import "OCItem.h"
 #import "OCSQLiteTransaction.h"
 
+
+#define INSTALL_TRANSACTION_ERROR_COLLECTION_RESULT_HANDLER \
+	__block NSError *transactionError = nil;  \
+	OCSQLiteDBResultHandler resultHandler = ^(OCSQLiteDB *db, NSError *error, OCSQLiteTransaction *transaction, OCSQLiteResultSet *resultSet) {  \
+		if (error != nil)  \
+		{  \
+			transactionError = error;  \
+		}  \
+	};
+
+
 @implementation OCDatabase (Schemas)
 
 #pragma mark - Schemas
@@ -87,13 +98,7 @@
 		upgradeMigrator:^(OCSQLiteDB *db, OCSQLiteTableSchema *schema, void (^completionHandler)(NSError *error)) {
 			// Migrate to version 2
 			[db executeTransaction:[OCSQLiteTransaction transactionWithBlock:^NSError *(OCSQLiteDB *db, OCSQLiteTransaction *transaction) {
-				__block NSError *transactionError = nil;
-				OCSQLiteDBResultHandler resultHandler = ^(OCSQLiteDB *db, NSError *error, OCSQLiteTransaction *transaction, OCSQLiteResultSet *resultSet) {
-					if (error != nil)
-					{
-						transactionError = error;
-					}
-				};
+				INSTALL_TRANSACTION_ERROR_COLLECTION_RESULT_HANDLER
 
 				// Add fileID column
 				[db executeQuery:[OCSQLiteQuery query:@"ALTER TABLE metaData ADD COLUMN fileID TEXT" resultHandler:resultHandler]];
@@ -176,13 +181,7 @@
 		upgradeMigrator:^(OCSQLiteDB *db, OCSQLiteTableSchema *schema, void (^completionHandler)(NSError *error)) {
 			// Migrate to version 3
 			[db executeTransaction:[OCSQLiteTransaction transactionWithBlock:^NSError *(OCSQLiteDB *db, OCSQLiteTransaction *transaction) {
-				__block NSError *transactionError = nil;
-				OCSQLiteDBResultHandler resultHandler = ^(OCSQLiteDB *db, NSError *error, OCSQLiteTransaction *transaction, OCSQLiteResultSet *resultSet) {
-					if (error != nil)
-					{
-						transactionError = error;
-					}
-				};
+				INSTALL_TRANSACTION_ERROR_COLLECTION_RESULT_HANDLER
 
 				// Add syncAnchor column
 				[db executeQuery:[OCSQLiteQuery query:@"ALTER TABLE metaData ADD COLUMN syncAnchor INTEGER" resultHandler:resultHandler]];
@@ -209,6 +208,7 @@
 				mdID : INTEGER	  	- unique ID used to uniquely identify and efficiently update a row
 				type : INTEGER    	- OCItemType value to indicate if this is a file or a collection/folder
 				syncAnchor: INTEGER	- sync anchor, a number that increases its value with every change to an entry. For files, higher sync anchor values indicate the file changed (incl. creation, content or meta data changes). For collections/folders, higher sync anchor values indicate the list of items in the collection/folder changed in a way not covered by file entries (i.e. rename, deletion, but not creation of files).
+				removed : INTEGER	- value indicating if this file or folder has been removed: 1 if it was, 0 if not (default). Removed entries are kept around until their delta to the latest syncAnchor value exceeds -[OCDatabase removedItemRetentionLength].
 				locallyModified: INTEGER- value indicating if this is a file that's been created or modified locally
 				localRelativePath: TEXT	- path of the local copy of the item, relative to the rootURL of the vault that stores it
 				path : TEXT	  	- full path of the item (e.g. "/example/file.txt")
@@ -217,12 +217,13 @@
 				fileID : TEXT		- OCFileID identifying the item
 				itemData : BLOB	  	- data of the serialized OCItem
 			*/
-			@"CREATE TABLE metaData (mdID INTEGER PRIMARY KEY, type INTEGER NOT NULL, syncAnchor INTEGER NOT NULL, locallyModified INTEGER NOT NULL, localRelativePath TEXT NULL, path TEXT NOT NULL, parentPath TEXT NOT NULL, name TEXT NOT NULL, fileID TEXT NOT NULL, itemData BLOB NOT NULL)",
+			@"CREATE TABLE metaData (mdID INTEGER PRIMARY KEY, type INTEGER NOT NULL, syncAnchor INTEGER NOT NULL, removed INTEGER NOT NULL, locallyModified INTEGER NOT NULL, localRelativePath TEXT NULL, path TEXT NOT NULL, parentPath TEXT NOT NULL, name TEXT NOT NULL, fileID TEXT NOT NULL, itemData BLOB NOT NULL)",
 
 			// Create indexes over path and parentPath
 			@"CREATE INDEX idx_metaData_path ON metaData (path)",
 			@"CREATE INDEX idx_metaData_parentPath ON metaData (parentPath)",
 			@"CREATE INDEX idx_metaData_synchAnchor ON metaData (syncAnchor)",
+			@"CREATE INDEX idx_metaData_removed ON metaData (removed)",
 		]
 		openStatements:@[
 			// Create trigger to delete thumbnails alongside metadata entries
@@ -231,9 +232,17 @@
 		upgradeMigrator:^(OCSQLiteDB *db, OCSQLiteTableSchema *schema, void (^completionHandler)(NSError *error)) {
 			// Migrate to version 4
 			[db executeTransaction:[OCSQLiteTransaction transactionWithBlock:^NSError *(OCSQLiteDB *db, OCSQLiteTransaction *transaction) {
-				__block NSError *transactionError = nil;
+				INSTALL_TRANSACTION_ERROR_COLLECTION_RESULT_HANDLER
 
-				// Delete existing metaData (as it lacks parentFileID info)
+				// Add "removed" column
+				[db executeQuery:[OCSQLiteQuery query:@"ALTER TABLE metaData ADD COLUMN removed INTEGER" resultHandler:resultHandler]];
+				if (transactionError != nil) { return(transactionError); }
+
+				// Create "removed" index
+				[db executeQuery:[OCSQLiteQuery query:@"CREATE INDEX idx_metaData_removed ON metaData (removed)" resultHandler:resultHandler]];
+				if (transactionError != nil) { return(transactionError); }
+
+				// Delete existing metaData (as it lacks parentFileID info, and versions of this schema < 4 serve only as cache)
 				[db executeQuery:[OCSQLiteQuery query:@"DELETE FROM metaData" resultHandler:^(OCSQLiteDB *db, NSError *error, OCSQLiteTransaction *transaction, OCSQLiteResultSet *resultSet) {
 					if (error != nil)
 					{
@@ -272,6 +281,44 @@
 		]
 		openStatements:nil
 		upgradeMigrator:nil]
+	];
+
+
+	// Version 2
+	[self.sqlDB addTableSchema:[OCSQLiteTableSchema
+		schemaWithTableName:OCDatabaseTableNameSyncJournal
+		version:2
+		creationQueries:@[
+			/*
+				recordID : INTEGER  		- unique ID used to uniquely identify and efficiently update a row
+				timestampDate : REAL		- NSDate.timeIntervalSinceReferenceDate at the time the record was added to the journal
+				inProgressSinceDate : REAL	- NSDate.timeIntervalSinceReferenceDate at the time the record was beginning to be processed
+				action : TEXT			- action to perform
+				path : TEXT			- path of the item targeted by the operation
+				recordData : BLOB		- archived OCSyncRecord data
+			*/
+			@"CREATE TABLE syncJournal (recordID INTEGER PRIMARY KEY, timestampDate REAL NOT NULL, inProgressSinceDate REAL, action TEXT NOT NULL, path TEXT NOT NULL, recordData BLOB)",
+		]
+		openStatements:nil
+		upgradeMigrator:^(OCSQLiteDB *db, OCSQLiteTableSchema *schema, void (^completionHandler)(NSError *error)) {
+			// Migrate to version 4
+			[db executeTransaction:[OCSQLiteTransaction transactionWithBlock:^NSError *(OCSQLiteDB *db, OCSQLiteTransaction *transaction) {
+				INSTALL_TRANSACTION_ERROR_COLLECTION_RESULT_HANDLER
+
+				// Drop unused V1 table
+				[db executeQuery:[OCSQLiteQuery query:@"DROP TABLE syncJournal" resultHandler:resultHandler]];
+				if (transactionError != nil) { return(transactionError); }
+
+				// Create it anew
+				[db executeQuery:[OCSQLiteQuery query:@"CREATE TABLE syncJournal (recordID INTEGER PRIMARY KEY, timestampDate REAL NOT NULL, inProgressSinceDate REAL, action TEXT NOT NULL, path TEXT NOT NULL, recordData BLOB)" resultHandler:resultHandler]];
+				if (transactionError != nil) { return(transactionError); }
+
+				return (transactionError);
+
+			} type:OCSQLiteTransactionTypeDeferred completionHandler:^(OCSQLiteDB *db, OCSQLiteTransaction *transaction, NSError *error) {
+				completionHandler(error);
+			}]];
+		}]
 	];
 }
